backend/requirements.txt
# FastAPI 核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic-settings==2.1.0

# 数据库相关
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# AI 相关 (Vanna 2.0)
vanna>=2.0.0
openai>=2.14.0
dashscope==1.14.1
pgvector==0.2.4

# Redis 缓存
redis[hiredis]==5.0.1

# 安全相关
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 数据处理
pandas==2.1.4
openpyxl==3.1.2
faker==22.0.0

# HTTP 客户端
httpx==0.25.2
aiohttp==3.9.1

# 环境变量管理
python-dotenv==1.0.0

# 日志和调试
loguru==0.7.2
